Added for consideration...
This is a Python-based end-user build script. There is no autogen/buildconf
step, nor a ./configure step. It uses apr/apu-config to fetch all the
right bits and simply Does The Work.

Still rather hacky/incomplete/demo, but it demonstrates a new, simple way
to skip the auto* mess (and leverage APR). It can build the library and
all the test programs, along with running the equiv of "make check".

* serfmake: new python building script


git-svn-id: http://serf.googlecode.com/svn/trunk@868 61a7d7f5-40b7-0310-9c16-bb0ea8cb1845